Read the MOs after the Overlap has been computed
